logout menu item is now a proper page (which redirects fast to login), closed #336
